\name{DBModelSelect-package}
\alias{DBModelSelect-package}
\alias{DBModelSelect}
\docType{package}
\title{
\packageTitle{DBModelSelect}
}
\description{
\packageDescription{DBModelSelect}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DBModelSelect}
\packageIndices{DBModelSelect}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DBModelSelect}

Maintainer: \packageMaintainer{DBModelSelect}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
