% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGLMSubsets.R
\name{FitGLMSubsets}
\alias{FitGLMSubsets}
\title{Perform all subsets regression for generalized linear models}
\usage{
FitGLMSubsets(
  response,
  data,
  family,
  intercept = TRUE,
  force_intercept = TRUE,
  ...
)
}
\arguments{
\item{response}{A character string specifying the name of the response variable.}

\item{data}{A dataframe containing a column corresponding to the response variable in addition to columns for each covariate of interest.}

\item{family}{A family suitable for supplying to the `glm` function specifying the error distribution and link function.}

\item{intercept}{A logical indicating whether an intercept term should be considered in models. Defaults to TRUE.}

\item{force_intercept}{A logical indicating whether to force an intercept term into all models if an intercept is desired. Defaults to TRUE.}

\item{...}{Additional arguments that may be supplied when calling `glm` to fit the models of interest.}
}
\value{
A list of fitted linear models suitable for use with the `StandICModelSelect` function.
}
\description{
Perform linear regression on all subsets of covariates supplied. Produces an output suitable for use with the `StandICModelSelect` function.
}
\examples{
# example code
# generate some data
data <- data.frame(s = rnorm(200), t = rnorm(200))
data$y <- data$s + rnorm(200)
# perform all subsets regression
model_list <- FitGLMSubsets(response = "y", data = data, family = gaussian(), intercept = TRUE, force_intercept = TRUE)
# perform model selection
model_select <- StandICModelSelect(model_list, IC = "AIC")
}
