% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandICPlot.R
\name{StandICPlot}
\alias{StandICPlot}
\title{Plot standardized information criteria}
\usage{
StandICPlot(model_select)
}
\arguments{
\item{model_select}{A list that was output using the `StandICModelSelect` function.}
}
\description{
Create a basic plot of the results of a model selection using standardized information criteria.
}
\examples{
# example code
# perform model selection
model_select <- StandICModelSelect(model_list, IC = "AIC")
# plot results of model selection
StandICPlot(model_select)
}
