\name{DBModelSelect-package}
\alias{DBModelSelect-package}
\alias{DBModelSelect}
\docType{package}
\title{
\packageTitle{DBModelSelect}
}
\description{
\packageDescription{DBModelSelect}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DBModelSelect}
\packageIndices{DBModelSelect}
The \code{DBModelSelect package} provides several methods of
model selection based in distributional theory. This includes
an implementation of selection using standardized information
criteria in the \code{StandICModelSelect} function, and
the implementation of an omnibus goodness-of-fit test for
linear models in the \code{BootGOFTestLM} function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shkoeneman/DBModelSelect}
}
}
\author{
\strong{Maintainer}: Scott H. Koeneman \email{Scott.Koeneman@jefferson.edu} 

}